separated spec files for number box
