+++ cypress/integration/v2/blocks/number-box/desktop-advanced.spec.js
@@ -1,0 +1,11 @@
+/**
+ * External dependencies
+ */
+import { registerTests } from '~stackable-e2e/helpers'
+
+/**
+ * Internal dependencies
+ */
+import { desktopAdvanced } from '.'
+
+describe( 'Number Box Block ( Desktop Advanced )', registerTests( [ desktopAdvanced ] ) )
